extract AsyncWrite out of RawHerderClient
